Introduce support for cloud provider registries
